requirements.txt
appdirs==1.4.3
-e git+https://github.com/kivy/buildozer.git@a8f764f7867f43711f84dc85a644308a4a9fd73f#egg=buildozer
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Cython==0.25.2
docutils==0.14
idna==2.6
Jinja2==2.10
Kivy==1.10.1.dev0
Kivy-Garden==0.1.4
MarkupSafe==1.0
pexpect==4.6.0
ptyprocess==0.5.2
pygame==1.9.3
Pygments==2.2.0
pyjnius==1.1.1
requests==2.18.4
sh==1.12.4
six==1.11.0
urllib3==1.22
virtualenv==16.0.0
